chore: rebase and fixes with main branch

+++ anomaly/delete-sheild.sh
@@ -8,6 +8,9 @@
 
 set -euo pipefail
 
+which curl > /dev/null || { echo >&2 "error: curl not found"; exit 3; }
+which jq > /dev/null || { echo >&2 "error: jq not found"; exit 3; }
+
 function usage() {
     cat <<END >&2
 USAGE: $0 [-e env] [-a access_token] [-i user_id] [-v|-h]
@@ -41,11 +44,8 @@ done
     exit 1
 }
 
-[[ -z ${access_token} ]] && {
-    echo >&2 -e "ERROR: no 'access_token' defined. \nopen -a safari https://manage.auth0.com/#/apis/ \nexport access_token=\`pbpaste\`"
-    exit 1
-}
-declare -r AUTH0_DOMAIN_URL=$(echo "${access_token}" | awk -F. '{print $2}' | base64 -di 2>/dev/null | jq -r '.iss')
+[[ -z ${access_token} ]] && { echo >&2 -e "ERROR: no 'access_token' defined. \nopen -a safari https://manage.auth0.com/#/apis/ \nexport access_token=\`pbpaste\`"; exit 1; }
+declare -r AUTH0_DOMAIN_URL=$(jq -Rr 'split(".") | .[1] | @base64d | fromjson | .iss' <<< "${access_token}")
 
 curl -s -X DELETE -H "Authorization: Bearer ${access_token}" -H 'content-type: application/json' \
     "${AUTH0_DOMAIN_URL}api/v2/anomaly/shields/${shield_id}"
